fixed job and scripts issue
fixed job and scripts issue
