+++ contrib/checkbox-ce-oem/checkbox-ce-oem-snap/series_classic22/snap/snapcraft.yaml
@@ -45,6 +45,7 @@ parts:
       - device-tree-compiler
       - linuxptp
       - snmp
+      - python3-rpyc
     override-prime: |
       snapcraftctl prime
       rm lib/systemd/system/alsa-utils.service
