+++ contrib/checkbox-ce-oem/checkbox-provider-ce-oem/bin/rpyc_server.py
@@ -3,54 +3,49 @@ import logging
 import os
 import rpyc
 import sys
-import time
 
 from contextlib import redirect_stdout, redirect_stderr
-from importlib import reload
 from importlib.machinery import SourceFileLoader
-from pathlib import Path
 from rpyc.utils.server import ThreadedServer
 
-from checkbox_support.scripts import (
-    run_watcher,
-    usb_read_write,
-)
 
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 handler_std = logging.StreamHandler(sys.stdout)
 
-
+CHECKBOX_PROVIDER_CEOEM_PATH = (
+    "/snap/checkbox-ce-oem/current/providers/checkbox-provider-ce-oem/bin"
+)
 LIBS = {
     "enable_serial_server": {
-        "source": (
-            "/snap/checkbox-ce-oem/current/providers/"
-            "checkbox-provider-ce-oem/bin/serial_test.py"
+        "source": os.path.join(
+            CHECKBOX_PROVIDER_CEOEM_PATH, "serial_test.py"
         ),
         "function": "server_mode",
     },
-    "configure_local_network": {
-        "source": (
-            "/snap/checkbox-ce-oem/current/providers/"
-            "checkbox-provider-ce-oem/bin/multiple_otg.py"
-        ),
-        "function": "configure_local_network",
-    },
     "serial_check": {
-        "source": __file__,
+        "source": os.path.join(
+            CHECKBOX_PROVIDER_CEOEM_PATH, "rpyc_test_methods.py"
+        ),
         "function": "serial_check",
     },
     "ethernet_check": {
-        "source": __file__,
+        "source": os.path.join(
+            CHECKBOX_PROVIDER_CEOEM_PATH, "rpyc_test_methods.py"
+        ),
         "function": "ethernet_check",
     },
     "configure_local_network": {
-        "source": __file__,
+        "source": os.path.join(
+            CHECKBOX_PROVIDER_CEOEM_PATH, "rpyc_test_methods.py"
+        ),
         "function": "configure_local_network",
     },
     "usb_storage_test": {
-        "source": __file__,
+        "source": os.path.join(
+            CHECKBOX_PROVIDER_CEOEM_PATH, "rpyc_test_methods.py"
+        ),
         "function": "usb_storage_test",
     },
 }
@@ -69,7 +64,8 @@ def capture_io_logs(func):
             try:
                 ret = func(*args, **kwargs)
             except SystemExit as exp:
-                ret = exp.code
+                logging.error(exp.code)
+                ret = None
 
         cls.logs = "stdout logs: {}".format(stdout.getvalue())
         cls.logs += "\nstderr logs: {}".format(stderr.getvalue())
@@ -98,69 +94,6 @@ def append_method_to_service(cls):
             setattr(cls, key, capture_io_logs(func))
 
     return cls
-
-
-# Method for testing
-def _get_ethernet_ifaces():
-    return set([i.name for i in Path("/sys/class/net").iterdir()])
-
-
-def _get_serial_ifaces():
-    return set([i.name for i in Path("/dev/serial/by-id").iterdir()])
-
-
-def _device_node_detect(func, device_type=""):
-    starting_ifaces = func()
-
-    attempts = 20
-    while attempts > 0:
-        now_ifaces = func()
-        # check if something disappeared
-        if not starting_ifaces == now_ifaces & starting_ifaces:
-            raise SystemExit(
-                "Interface(s) disappeared: {}".format(
-                    ", ".join(list(starting_ifaces - now_ifaces))
-                )
-            )
-        new_ifaces = now_ifaces - starting_ifaces
-        if new_ifaces:
-            print()
-            print(
-                "New interface(s) detected: {}".format(
-                    ", ".join(list(new_ifaces))
-                )
-            )
-            return new_ifaces
-        time.sleep(1)
-        print(".", end="", flush=True)
-        attempts -= 1
-    print()
-    raise SystemExit(
-        "Failed to detect new {} interface".format(device_type)
-    )
-
-
-def serial_check():
-    return _device_node_detect(_get_serial_ifaces, "serial")
-
-
-def ethernet_check():
-    return _device_node_detect(_get_ethernet_ifaces, "ethernet")
-
-
-def usb_storage_test(usb_type):
-    logger.info("%s drive read/write testing on RPYC", usb_type)
-    usb_read_write.REPETITION_NUM = 2
-    watcher = run_watcher.USBStorage(usb_type)
-    mounted_partition = watcher.run_insertion()
-    logger.info(
-        "%s drive been mounted to %s on RPYC", usb_type, mounted_partition
-    )
-    watcher.run_storage(mounted_partition)
-    # usb_read_write been imported in run_watcher
-    # the temporary mount point been created while import it and been removed in gen_random_file function
-    # thus, we need to reload it to create temporary mount point in every testing cycle
-    reload(usb_read_write)
 
 
 class RpycTestService(rpyc.Service):
